Fix JsonUtils.TryLoad collection issue
Updated the JsonUtils class in the Kivikko.Json library to modify the behavior of the TryLoad method. It was transformed to no longer instantiate a new object when the file does not exist or the deserialization fails. Instead, it returns the 'default' value for the type argument. A SaveLoadTest in JsonUtilsTests was also added to ensure validity and functionality.